bugfix: 修改逻辑集群业务逻辑:适配行内升级方案

See merge request bigdata-cloud/knowsearch!369

+++ arius-admin/arius-admin-biz/src/main/java/com/didichuxing/datachannel/arius/admin/biz/cluster/impl/ClusterPhyManagerImpl.java
@@ -383,7 +383,8 @@ public class ClusterPhyManagerImpl implements ClusterPhyManager {
             return Result.buildParamIllegal("集群资源类型非法");
         }
         List<String> clusters = Lists.newArrayList();
-        ClusterLogic clusterLogic = clusterLogicService.getClusterLogicById(clusterLogicId);
+        ClusterLogic clusterLogic =
+                clusterLogicService.getClusterLogicById(clusterLogicId ).stream().findFirst().orElse(null);
         if (clusterLogic == null) {
             return Result.buildFail("选定的逻辑集群不存在");
         }
@@ -1092,7 +1093,8 @@ public class ClusterPhyManagerImpl implements ClusterPhyManager {
         }
         Result<List<String>> canBeAssociatedClustersPhyNames = Result.buildSucc(Lists.newArrayList());
         if (clusterLogicId != null) {
-            ClusterLogic clusterLogicById = clusterLogicService.getClusterLogicById(clusterLogicId);
+            ClusterLogic clusterLogicById =
+                    clusterLogicService.getClusterLogicById(clusterLogicId).stream().findFirst().orElse(null);
             if (clusterLogicById == null) {
                 return Result.buildFail("选定的逻辑集群不存在");
             }
@@ -1458,7 +1460,11 @@ public class ClusterPhyManagerImpl implements ClusterPhyManager {
 
         List<Long> clusterLogicIds = clusterPhyContext.getAssociatedClusterLogicIds();
         for (Long clusterLogicId : clusterLogicIds) {
-            final ClusterLogic deleteById = clusterLogicService.getClusterLogicById(clusterLogicId);
+            final ClusterLogic deleteById = clusterLogicService.getClusterLogicById(clusterLogicId)
+                    .stream().findFirst().orElse(null);
+            if (Objects.isNull(deleteById)){
+                continue;
+            }
             Result<Void> deleteLogicClusterResult = clusterLogicService.deleteClusterLogicById(clusterLogicId, operator,
                 projectId);
             if (deleteLogicClusterResult.failed()) {
